+++ example/lib/ui/MainActivity.dart
@@ -52,7 +52,7 @@ class _MainActivityState extends State<MainActivity>
     // TODO: implement initState
 
     super.initState();
-    _tabController = TabController(length: 5, vsync: this);
+    _tabController = TabController(length: 6, vsync: this);
     _menuController = MenuController(vsync: this,direction: ScrollDirection.LEFT);
     _pageController = PageController(initialPage: 1);
     views = [
@@ -88,6 +88,7 @@ class _MainActivityState extends State<MainActivity>
               Tab(child: Text("嵌套滚动视图")),
               Tab(child: Text("动态变化指示器+Navigator")),
               Tab(child: Text("主动刷新")),
+              Tab(child: Text("四个方向不同风格测试绘制")),
             ],
             controller: _tabController,
           ):null,